requirements.txt
pytest>=7.4.0
allure-pytest>=2.13.2
selene>=2.0.b4
selenium>=4.11.2